Extract digitalcomponent
