+++ lib/ui/widgets/digital.js
@@ -3,57 +3,18 @@ var React = require('react');
 var BaseWidget = require('./basewidget');
 require('util').inherits(Digital, BaseWidget);
 
+var componentClass = require('./digitalcomponent');
+
 function Digital(id, options, streamBundle, instrumentPanel) {
   BaseWidget.call(this, id, options, streamBundle, instrumentPanel);
   this.options = options;
-  this.widget = React.createElement(this.createClass(this.displayValue, this.render),{
+  this.widget = React.createElement(componentClass,{
     key: id,
     unit: this.getUnitForPath(options.path) || "",
-    value: streamBundle.getStreamForSourcePath(options.sourceId, options.path),
+    valueStream: streamBundle.getStreamForSourcePath(options.sourceId, options.path).toProperty(),
     label: this.getLabelForPath(options.path),
     sourceId: options.sourceId
   });
-}
-
-Digital.prototype.createClass = function(displayValue, render) {
-  return React.createClass({
-    mixins: [require('../../util/baconvaluewithlabelmixin')],
-    displayValue: displayValue,
-    render: render
-  })
-}
-
-Digital.prototype.render = function() {
-  return (
-    <svg height="100%" width="100%" viewBox="0 0 20 40">
-      <text x="10" y="8" textAnchor="middle" fontSize="5" dominant-baseline="middle">
-        {this.state.label + " " + this.props.sourceId}
-      </text>
-      <g transform="translate(10, 0)">
-        <text textAnchor="middle" y="32" fontSize="26" dominant-baseline="middle">
-          {this.state.value.toString()}
-        </text>
-        <text textAnchor="end" x="40" y="32" fontSize="8" dominant-baselie="middle">
-          {this.state.unit}
-        </text>
-      </g>
-    </svg>
-  )
-}
-
-Digital.prototype.displayValue = function(value) {
-  if(typeof value === 'number') {
-    var v = Math.abs(value);
-    if(v >= 50) {
-      return value.toFixed(0);
-    } else if(v >= 10) {
-      return value.toFixed(1);
-    } else {
-      return value.toFixed(2);
-    }
-  }
-
-  return value;
 }
 
 Digital.prototype.getReactElement = function() {
